nrf_802154: rev e658759beb67c8060664ecca9d2f79ac5ff8e8f7
This commit updates revision of the nrf_802154 component.
