+++ nrf_802154/driver/src/nrf_802154_peripherals.h
@@ -65,7 +65,7 @@ extern "C" {
  *
  */
 #ifndef NRF_802154_HIGH_PRECISION_TIMER_INSTANCE_NO
-#define NRF_802154_HIGH_PRECISION_TIMER_INSTANCE_NO 0
+#define NRF_802154_HIGH_PRECISION_TIMER_INSTANCE_NO 1
 #endif
 
 /**
@@ -84,7 +84,7 @@ extern "C" {
  *
  */
 #ifndef NRF_802154_TIMER_INSTANCE_NO
-#define NRF_802154_TIMER_INSTANCE_NO 1
+#define NRF_802154_TIMER_INSTANCE_NO 0
 #endif
 
 /**
