+++ nrf_802154/driver/src/nrf_802154_trx.c
@@ -229,6 +229,7 @@ static void * volatile mp_receive_buffer;
 /** Initialize TIMER peripheral used by the driver. */
 void nrf_timer_init(void)
 {
+    nrf_timer_task_trigger(NRF_802154_TIMER_INSTANCE, NRF_TIMER_TASK_SHUTDOWN);
     nrf_timer_mode_set(NRF_802154_TIMER_INSTANCE, NRF_TIMER_MODE_TIMER);
     nrf_timer_bit_width_set(NRF_802154_TIMER_INSTANCE, NRF_TIMER_BIT_WIDTH_32);
     nrf_timer_frequency_set(NRF_802154_TIMER_INSTANCE, NRF_TIMER_FREQ_1MHz);
@@ -472,7 +473,6 @@ void nrf_802154_trx_init(void)
 
     nrf_802154_trx_module_reset();
 
-    nrf_timer_init();
 #if defined(RADIO_INTENSET_SYNC_Msk)
     nrf_802154_swi_init();
 #endif
@@ -486,6 +486,7 @@ void nrf_802154_trx_enable(void)
 
     assert(m_trx_state == TRX_STATE_DISABLED);
 
+    nrf_timer_init();
     nrf_radio_reset();
 
 #if defined(NRF52840_XXAA) || \
