+++ nrf_802154/sl/include/nrf_802154_sl_periphs.h
@@ -127,7 +127,7 @@
  *
  */
 #ifndef NRF_802154_HIGH_PRECISION_TIMER_INSTANCE_NO
-#define NRF_802154_HIGH_PRECISION_TIMER_INSTANCE_NO 0
+#define NRF_802154_HIGH_PRECISION_TIMER_INSTANCE_NO 1
 #endif
 
 /**
