+++ nrf_802154/sl/sl_opensource/include/protocol/mpsl_fem_protocol_api.h
@@ -308,21 +308,17 @@ void mpsl_fem_cleanup(void);
 /** @brief Splits transmit power value into components to be applied on each stage on transmit path.
  *
  * @note If the exact value of @p power cannot be achieved, this function attempts to use less
- * power to not exceed constraint. However, if @p power is lower than the minimum achievable power,
- * or larger than the maximum achievable power, the function returns failure.
+ * power to not exceed constraint.
  *
  * @param[in]  power            TX power requested for transmission on air.
  * @param[out] p_tx_power_split Components of tx_power to be applied for stages on transmit path.
- *
- * @retval  0               Calculation performed successfully.
- * @retval  - ::NRF_EINVAL  Given @p power cannot be achieved. If requested value is too high
- *                          the @p p_tx_power_split will be set to a value representing maximum
- *                          achievable power. If the requested value is too low, the
- *                          @p p_tx_power_split will be set to a value representing minimum
- *                          achievable power.
+ *                              If requested @p power is too high, the split will be set to
+ *                              a value representing maximum achievable power. If the requested
+ *                              @p power is too low, the split will be set to a value representing
+ *                              minimum achievable power.
  */
-int32_t mpsl_fem_tx_power_split(const mpsl_tx_power_t power,
-                                mpsl_tx_power_split_t *const p_tx_power_split);
+void mpsl_fem_tx_power_split(const mpsl_tx_power_t power,
+                             mpsl_tx_power_split_t *const p_tx_power_split);
 
 /** @brief Sets PA gain.
  *
